+++ Assets/JI/00.Scripts/00.Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3d43e6de8f8c3df41bf5eb2e902c2980
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
