+++ Assets/JI/00.Scripts/00.Player/JI_ResourceController.cs.meta
