+++ Assets/JI/00.Scripts/01.Manager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4f2e92ff9203ee94b89616e30ed11e2e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
