+++ Assets/JI/00.Scripts/01.Manager/JI_CustomManager.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 812a94fe953b6164fa96f925a6c2ff81
+guid: 6c558d2b0dbbaef4ab389f441fab3a56
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
