+++ Assets/JI/00.Scripts/01.Manager/JI_GM.cs
@@ -1,0 +1,13 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class JI_GM : MonoBehaviour
+{
+   
+    public void OnClikToStat()
+    {
+        SceneManager.LoadScene("JI_Scene");
+    }
+}
