+++ ProjectSettings/EditorBuildSettings.asset
@@ -5,7 +5,7 @@ EditorBuildSettings:
   m_ObjectHideFlags: 0
   serializedVersion: 2
   m_Scenes:
-  - enabled: 1
+  - enabled: 0
     path: Assets/Scenes/SampleScene.unity
     guid: 2cda990e2423bbf4892e6590ba056729
   - enabled: 1
@@ -17,4 +17,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/BAEK_C/Scenes/StageScene.unity
     guid: 1d145c9c8c2c36941a0e385f01f2e435
+  - enabled: 1
+    path: Assets/JI/03.Scene/JI_Scene.unity
+    guid: d946d44f0fab9f6448874201f1d54b8e
   m_configObjects: {}
